Neighborhood Map
The neighborhood tour application is complex enough and incorporates a
variety of data points that it can easily become unwieldy to manage.
There are a number of frameworks, libraries and APIs available to make
this process more manageable and many employers are looking for
specific skills in using these packages.